requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
httpx==0.26.0
pydantic==2.5.3
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
python-dotenv==1.0.0
